whatsapp-mcp-medgemma-project/requirements.txt
# --- Core / Utilities ---
absl-py
altair
annotated-types
anyio
astunparse
attrs
blinker
cachetools
certifi
chardet
charset-normalizer
click
colorama
coloredlogs
contourpy
cycler
filelock
flatbuffers
fonttools
fsspec
gast
gitdb
GitPython
google-pasta
grpcio
h11
h5py
httpcore
httpx
humanfriendly
idna
itsdangerous
Jinja2
joblib
jsonschema
jsonschema-specifications
keras
kiwisolver
libclang
lightning-utilities
Markdown
markdown-it-py
MarkupSafe
mdurl
ml_dtypes
mpmath
namex
narwhals
networkx
ngrok
opt_einsum
optree
packaging
protobuf
py-cpuinfo
pyarrow
pydantic
pydantic_core
pydeck
Pygments
pyparsing
pyreadline3
python-dateutil
pytz
PyYAML
referencing
rich
rpds-py
scipy
seaborn
six
smmap
sniffio
sympy
tenacity
tensorboard
tensorboard-data-server
termcolor
threadpoolctl
toml
tornado
tqdm
typing-inspection
typing_extensions
tzdata
ultralytics
ultralytics-thop
urllib3
watchdog

# --- Web servers / frameworks ---
flask
fastapi
uvicorn
python-dotenv

# --- HTTP requests ---
requests

# --- ML / PyTorch ---
torch
torchvision
torchmetrics
scikit-learn
numpy

# --- Image handling ---
pillow
opencv-python

# --- Visualization / debugging ---
matplotlib
psutil
streamlit

# --- PDF reports ---
reportlab

# --- Ollama Python SDK ---
ollama

# --- Extra ---
Werkzeug
starlette
